backend/requirements.txt
pymongo[srv]==3.12
python-dotenv
fastapi
google-auth
google-auth-oauthlib
uvicorn
python-jose